dev-requirements.txt
# Testing
pytest==8.4.2
pytest-django==4.11.1
factory_boy==3.3.3
Faker==37.11.0
attrs==25.4.0

# Code quality & formatting
black==23.9.1
flake8==7.3.0
flake8-bugbear==23.9.16
isort==7.0.0

# Type checking
mypy==1.18.2
django-stubs==5.2.7
django-stubs-ext==5.2.7
types-PyYAML==6.0.12.20250915

# Pre-commit
pre_commit==4.3.0
cfgv==3.4.0
identify==2.6.15
nodeenv==1.9.1
virtualenv==20.35.3

# Dependencies
click==8.3.0
distlib==0.4.0
filelock==3.20.0
iniconfig==2.3.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyYAML==6.0.3
typing_extensions==4.15.0
